% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icdData-package.R
\docType{package}
\name{icd-package}
\alias{icd-package}
\alias{package-icdData}
\title{Data and pre-processing functions to support the 'icd' package.}
\description{
Data and pre-processing functions to support the 'icd' package.
}
\author{
Jack O. Wasey \email{jack@jackwasey.com}
}
\concept{
ICD-9 ICD-10 comorbidity comorbidities
}
\keyword{data}
\keyword{misc}
\keyword{utilities}

